backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.2.1
openai==2.8.0
gunicorn==23.0.0

